Fix printf warning on 64bit

+++ x.cpp
@@ -61,7 +61,7 @@ void test_array()
 
    ttl::array<int, 3> i3 = { {1,2,3} };
    for (ttl::array<int, 3>::const_iterator i = i3.begin(); i != i3.end(); ++i)
-      printf("%d/%d: %d\n", i - i3.begin(), i3.size(), *i);
+      printf("%ld/%d: %d\n", long(i - i3.begin()), i3.size(), *i);
    ttl::array<int, 3> i30;
    i30 = i3;
    printf("equal? %s\n", i3 == i30 ? "yes": "no");
